refactor and add clip example
