+++ smse/models/__init__.py
@@ -3,7 +3,10 @@ from smse.logging import get_logger
 logger = get_logger(__name__)
 
 try:
-    from .st_models import SentenceTransformerCLIPModel, SentenceTransformerTextModel
+    from .sentence_transformers import (
+        SentenceTransformerCLIPModel,
+        SentenceTransformerTextModel,
+    )
 except ImportError:
     logger.warning(
         "SentenceTransformer is not installed. "
@@ -12,7 +15,7 @@ except ImportError:
     )
 
 try:
-    from .imagebind_models import ImageBindModel
+    from .imagebind import ImageBindModel
 except ImportError:
     logger.warning(
         "ImageBind is not installed. "
